server: GET /v1/policies - Don't return partial modules
Previously we would iterate over all modules on the compiler and then
look them up by id in the store. This causes problems when there are
partial compilation results on the compiler that have not been put in
the store.

To correct this we will now reverse how it iterates, instead iterating
over a list of policies from the store and looking up their AST from
the compiler as needed.

Fixes: #2036
Signed-off-by: Patrick East <east.patrick@gmail.com>