topdown: Improve inlining controls
This commit improves the inlining controls in two ways:

1. The disable parameter is supposed to specify a prefix to disable
inlining on (not an exact path). This is important because otherwise
callers have to enumerate every single path under the root which can
be impractical in some cases. This commit updates the check to be symmetric.

2. The disable parameter should apply to base documents as well--in
some cases we need to ensure that documents can be referenced in the
result of PE (e.g., if a transitive dependent of an entrypoint/hotspot
is preserved after a build, any documents referred to in the
transitive dependent must be available otherwise the result of a build
is not semantically equivalent.) This commit updates the evaluator to
check the inlining control before resolving base refs.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>