docs: Add simple sorting to ecosystem
Previously the list of integrations did not give new users
any sense of how popular or mature the integrations are.  It
showed them in alphabetical order.

This change implements a simple sorting order based on
the amount of content for each integration.  Content
corresponds very roughly to popularity/maturity.

This implementation is static.  In the future the user should
be able to sort and filter the content dynamically.

Signed-off-by: Tim Hinrichs <tim@styra.com>