topdown: Fix copy propagation to treat input and data as safe
In a7dd95ad8f0a91ab1b89904be654c3d29088c83c we fixed copy propagation
for certain kinds of negated expressions however the fix contained an
issue wherein reserved vars were not marked as safe. This meant that
expressions may not be added back during post-processing. Since the
result was not checked for safety, it meant that unsafe results could
be generated. This commit fixes copy propagation to treat reserved
vars as safe, use the same safety check visitor params as the compiler
(for consistency), and adds a check for unsafe outputs--in that case,
the result is just discarded (which is better than generating an error
since we can't test that code path.)

Signed-off-by: Torin Sandall <torinsandall@gmail.com>