docs: Update envoy tutorial to use health probes
Previously the envoy tutorial would leave the default OPA
address (":8181") and didn't configure health probes. This meant
that the OPA API would be potentially exposed to other clients
in the cluster, and that kubernetes didn't wait on OPA to startup
to declare the pod as ready.

The guide now configures the main API on localhost and sets a
diagnostic address that is used with liveness and ready probes.

This does require a change in the proxy_init to whitelist the
diagnostic port for OPA to ensure that requests bypass the envoy
proxy.

Signed-off-by: Patrick East <east.patrick@gmail.com>