wasm: Improve makefile to use wildcards and pattern substitution
As more built-ins get implemented we will want to split them into
separate files. This change just updates the makefile so source files
don't have to listed explicitly. This makes it easier to maintain the
Makefile because we don't have to audit every single file to make sure
it's added correctly. This will also make it easier to do auto deps.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>