topdown: Disable copy propagation when shallow inlining enabled
When shallow inlining is enabled do not perform aggressive inlining
operations like copy propagation or negation inlining. We want the
output of PE to be as structurally similar to the original as possible
during shallow inlining and these other optimizations are opposed to
that--they attempt to pre-evaluate as much of the policy as possible.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>