topdown: Fix for empty documents generated by partial rules
This commit fixes an issue in partial eval whereby OPA would generate
an incorrect expression referring to a non-existent/undefined virtual
document. This would occur if an expression referenced the full extent
of a virtual document generated by partial rules and those partial
rules were all undefined. In that case, the reference needs to be
unified with the default (empty) value for that virtual document.

Fixes #2492

Signed-off-by: Torin Sandall <torinsandall@gmail.com>