wasm: Update to llvm 10 build tools
As part of this update we needed to change the memory grow builtins.
The version we used were removed in https://reviews.llvm.org/D56645
after (apparently) having been deprecated for some time. The newer
versions match up with more recent versions of the WASM spec (see
https://github.com/WebAssembly/spec/issues/627 for more context).

The WebAssembly/wabt version is also updated to the latest release to
correct build issues with the older one and llvm 10.

Upgrading should prevent a deadlock we saw periodically in the CI
while building the OPA wasm binaries.

Signed-off-by: Patrick East <east.patrick@gmail.com>