Add periodic fuzzer job and required tooling
This adds the tooling from:

https://github.com/tsandall/fuzz-opa

plus some new helper scripts and make targets to use it as a pass/fail
CI step.

The script will (as of now) run the fuzzing for an hour and raise an
error if any crashers were found. We can adjust the timing as needed,
this initial setting is pretty arbitrary.

Signed-off-by: Patrick East <east.patrick@gmail.com>