ast: Deprecate re_match in favour of regex.match
This ought to have been done a while ago. Going forward we prefer all
built-in functions to be namespaced. We made that change after
introducing `re_match` but before introducing several other
regex-related functions. This commit deprecates `re_match` in a
backwards-compatible manner.

Also, move a few built-in declarations up in the ast/builtins.go file
because they were in the wrong sections.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>