Allow input as PEM strings to x509 built-ins
In the PR discussion here: https://github.com/open-policy-agent/opa/pull/2564#issuecomment-662012650
we decided that PEM string input was also a valid use case to support.

This change allows string PEM data to be passed as input to both cert
and csr functions.

It also moves the cert and csr functions to be co-located.

Signed-off-by: Charlie Egan <charlieegan3@users.noreply.github.com>