topdown: Fix function output handling during PE
PE was not handling function output correctly when the output was not
unified with a variable. The output would essentially be ignored. This
was fine in all cases except when the function output could be
`false`. In that case, the `false` value would never be tested.

This commit fixes the issue by checking if the function output is
unified. If the output is not unified, PE saves an expression
containing the function output value so that the saved query tests for
`false` values.

Fixes #2573

Signed-off-by: Torin Sandall <torinsandall@gmail.com>